rename bayesian blocks files; fix doctests
