+++ astropy/stats/histogram.py
